spec: Require shared libraries with architecture-specific dependencies
This is an architecture-specific package (ie., not noarch), and the
json-c and libdnf packages only provide the architecture-specific
libjson-c.so and libdnf.so shared libraries.  According to the Fedora
packaging guidelines [1], when explicit library dependencies are needed,
they SHOULD be architecture-specific unless the packages are noarch.

Fallout from 8dddb545597e31a45304bec0d33e89e52a8fa705

[1] https://docs.fedoraproject.org/en-US/packaging-guidelines/